Code clean up.

+++ workspace-manager-clienttests/src/main/java/scripts/disruptivescripts/DeleteInitialPods.java
@@ -12,8 +12,10 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
 public class DeleteInitialPods extends DisruptiveScript {
     private static final Logger logger = LoggerFactory.getLogger(DeleteInitialPods.class);
@@ -23,7 +25,7 @@ public class DeleteInitialPods extends DisruptiveScript {
         manipulatesKubernetes = true;
     }
 
-    protected static final int secondsToWaitBeforeStartingDisrupt = 15;
+    protected static final int SECONDS_TO_WAIT_BEFORE_STARTING_DISRUPT = 15;
 
     @Override
     public void setParameters(List<String> parameters) {
@@ -36,37 +38,34 @@ public class DeleteInitialPods extends DisruptiveScript {
     @Override
     public void disrupt(List<TestUserSpecification> testUsers) throws Exception {
         // give user journey threads time to get started before disruption
-        TimeUnit.SECONDS.sleep(secondsToWaitBeforeStartingDisrupt);
+        TimeUnit.SECONDS.sleep(SECONDS_TO_WAIT_BEFORE_STARTING_DISRUPT);
 
         logger.info("Starting disruption - all initially created api pods will be deleted, one by one.");
 
-        /*List<V1Deployment> allDeployments = KubernetesClientUtils.listDeployments();
-        if (allDeployments.size() == 0) {
-            throw new RuntimeException("No deployment not found.");
-        }*/
-
         String componentLabelKey = "app.kubernetes.io/component";
         String componentLabelVal = "workspacemanager";
-        V1Deployment workspacemanagerDeployment = getDeployment(componentLabelKey, componentLabelVal);
+        V1Deployment workspacemanagerDeployment = getDeployment(componentLabelKey, componentLabelVal).orElse(null);
         if (workspacemanagerDeployment == null) {
             throw new RuntimeException("WorkspaceManager deployment not found.");
         }
 
-        List<String> podsToDelete = new ArrayList<>();
-        KubernetesClientUtils.listPods().stream()
+        List<String> podsToDelete = KubernetesClientUtils.listPods().stream()
                 .filter(
                         pod ->
                                 pod.getMetadata().getLabels().containsKey("app.kubernetes.io/component")
                                         && pod.getMetadata().getLabels().get("app.kubernetes.io/component").equals("workspacemanager"))
-                .forEach(p -> podsToDelete.add(p.getMetadata().getName()));
+                .map(pod -> pod.getMetadata().getName())
+                .collect(Collectors.toList());
 
         // delete original pods, and give them a chance to recover
         for (String podName : podsToDelete) {
             logger.debug("delete pod: {}", podName);
-            workspacemanagerDeployment = getDeployment(componentLabelKey, componentLabelVal);
-            KubernetesClientUtils.printApiPods(workspacemanagerDeployment);
-            KubernetesClientUtils.deletePod(podName);
-            KubernetesClientUtils.waitForReplicaSetSizeChange(workspacemanagerDeployment, podsToDelete.size());
+            workspacemanagerDeployment = getDeployment(componentLabelKey, componentLabelVal).orElse(null);
+            if (workspacemanagerDeployment != null) {
+                KubernetesClientUtils.printApiPods(workspacemanagerDeployment);
+                KubernetesClientUtils.deletePod(podName);
+                KubernetesClientUtils.waitForReplicaSetSizeChange(workspacemanagerDeployment, podsToDelete.size());
+            }
         }
 
         logger.debug("original pods:");
@@ -74,13 +73,12 @@ public class DeleteInitialPods extends DisruptiveScript {
         KubernetesClientUtils.printApiPods(workspacemanagerDeployment);
     }
 
-    private V1Deployment getDeployment(String labelKey, String labelVal) throws ApiException {
+    private Optional<V1Deployment> getDeployment(String labelKey, String labelVal) throws ApiException {
         return KubernetesClientUtils.listDeployments().stream()
                 .filter(
                         deployment -> deployment.getMetadata().getLabels().containsKey(labelKey)
                                 && deployment.getMetadata().getLabels().get(labelKey).equals(labelVal)
                 )
-                .findFirst()
-                .orElse(null);
+                .findFirst();
     }
 }
